+++ nutzboot-demo/nutzboot-demo-simple/nutzboot-demo-simple-thymeleaf/src/main/java/io/nutz/demo/simple/MainLauncher.java
@@ -1,21 +1,22 @@
-package io.nutz.demo.dubbo.rpc;
+package io.nutz.demo.simple;
 
 import org.nutz.boot.NbApp;
 import org.nutz.ioc.loader.annotation.IocBean;
+import org.nutz.lang.util.NutMap;
 import org.nutz.mvc.annotation.At;
 import org.nutz.mvc.annotation.Ok;
 
 @IocBean
 public class MainLauncher {
-    
-    @Ok("raw")
-    @At("/time/now")
-    public long now() {
-        return System.currentTimeMillis();
+
+    @At({"/", "/index"})
+    @Ok("th:/index.html")
+    public NutMap index() {
+        return NutMap.NEW().setv("name", "NB").setv("age", 18);
     }
 
     public static void main(String[] args) throws Exception {
-        new NbApp(MainLauncher.class).run();
+        new NbApp().run();
     }
 
 }