+++ nutzboot-starter-nutz-mvc/src/main/java/org/nutz/boot/starter/nutz/mvc/NbMvcLoading.java
@@ -2,13 +2,20 @@ package org.nutz.boot.starter.nutz.mvc;
 
 import org.nutz.boot.AppContext;
 import org.nutz.ioc.Ioc;
+import org.nutz.lang.Stopwatch;
+import org.nutz.log.Log;
+import org.nutz.log.Logs;
+import org.nutz.mvc.LoadingException;
 import org.nutz.mvc.Mvcs;
 import org.nutz.mvc.NutConfig;
+import org.nutz.mvc.Setup;
 import org.nutz.mvc.UrlMapping;
 import org.nutz.mvc.annotation.Localization;
 import org.nutz.mvc.impl.NutLoading;
 
 public class NbMvcLoading extends NutLoading {
+
+    private static final Log log = Logs.get();
     
     public UrlMapping load(NutConfig config) {
         config.setMainModule(AppContext.getDefault().getMainClass());
@@ -26,5 +33,26 @@ public class NbMvcLoading extends NutLoading {
             super.evalLocalization(config, mainModule);
         else
             super.evalLocalization(config, NbMainModule.class);
+    }
+    
+    public void depose(NutConfig config) {
+        if (log.isInfoEnabled())
+            log.infof("Nutz.Mvc[%s] is deposing ...", config.getAppName());
+        Stopwatch sw = Stopwatch.begin();
+
+        // Firstly, upload the user customized desctroy
+        try {
+            Setup setup = config.getAttributeAs(Setup.class, Setup.class.getName());
+            if (null != setup)
+                setup.destroy(config);
+        }
+        catch (Exception e) {
+            throw new LoadingException(e);
+        }
+
+        // Done, print info
+        sw.stop();
+        if (log.isInfoEnabled())
+            log.infof("Nutz.Mvc[%s] is down in %sms", config.getAppName(), sw.getDuration());
     }
 }