+++ ChangeLog.md
@@ -1,5 +1,16 @@
 # NB进化史
 
+## 2.0-RC2 "?"
+
+* 变更:
+	* add: new NbApp()可以不传类名,从堆栈自动推断
+	* add: starter-mongodb by @qingerg
+	* add: starter-tomcat by @ben
+	* add: starter-beetlsql 来自beetl的SQL解决方案
+	* add: starter-sharding-jdbc 分库分表
+	* add: starter-thymeleaf 模板引擎
+	* fix: jetty扫描websocket的endpoint有问题
+
 ## 2.0-RC "属于"
 
 首先,必须高亮一下[@qinerg](https://github.com/qinerg)提交的[starter-undertow](https://gitee.com/nutz/nutzboot/tree/dev/nutzboot-starter-undertow),嗷嗷嗷. Jetty与Undertow任意切换^_^